Move getNewInfo test to main file
